third_party/WebKit/LayoutTests/typedcssom/inlinestyle/transform.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

test(function() {
  var transform = new CSSTransformValue([new CSSRotation(new CSSAngleValue(20, 'deg'))]);

  assert_equals(testElement.styleMap.get('transform'), null);

  testElement.styleMap.set('transform', transform);
  assert_equals(testElement.style.transform, 'rotate(20deg)');
}, "Setting transform works on a regular element.");

test(function() {
  var transform = new CSSTransformValue([new CSSRotation(new CSSAngleValue(30, 'deg'))]);

  testElement.styleMap.set('transform', transform);
  assert_equals(testElement.styleMap.get('transform').cssText, 'rotate(30deg)');
}, "Getting transform works on a regular element.");

test(function() {
  testElement.style.transform = 'translateY(50px)';
  testElement.styleMap.entries();
}, "Unhandled case doesn't crash.");

</script>
